requirements.txt
opencv-python==4.8.0.74
mediapipe==0.10.8
pyserial==3.5
numpy==1.24.3
